Implemented Requested Changes All kafka stream beans moved to KafkaStreamConfig Removed FileStatusDto and FileStatusMapper createEntry method added to FileStatusService
