+++ service/pom.xml
@@ -19,7 +19,6 @@
         <kafka-streams.version>3.1.0</kafka-streams.version>
         <awaitility.version>4.2.0</awaitility.version>
         <spring.cloud.version>2021.0.2</spring.cloud.version>
-        <property name="hibernate.enable_lazy_load_no_trans" value="true"/>
     </properties>
     <dependencyManagement>
         <dependencies>
