+++ service/src/main/java/com/bankbazaar/kafka/service/producer/KafkaProducer.java
@@ -1,18 +1,19 @@
 package com.bankbazaar.kafka.service.producer;
 
 import com.bankbazaar.kafka.core.model.Data;
-import com.bankbazaar.kafka.dto.model.FileStatusDto;
+import com.bankbazaar.kafka.core.model.FileStatusEntity;
 import com.bankbazaar.kafka.service.mapper.DataMapper;
 import com.bankbazaar.kafka.dto.model.DataDto;
 import com.bankbazaar.kafka.service.model.Response;
 import com.bankbazaar.kafka.service.service.FileStatusService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
 
-@Service
+@Configuration
 public class KafkaProducer {
     private  static  final  String TOPIC = "Create_CSV";
 
@@ -26,16 +27,9 @@ public class KafkaProducer {
 
     public Response sendData(DataDto data)
     {
-        FileStatusDto fileData = new FileStatusDto();
-        fileData.setFileName(data.getFileName());
-        fileData.setStatus("NEW");
-        FileStatusDto response = fileStatusService.insert(fileData);
+        FileStatusEntity response = fileStatusService.createEntry(data);
         data.setId(response.getId());
         this.kafkaTemplate.send(TOPIC,mapper.dtoToDomain(data));
-
-        FileStatusDto controllerResponse = fileStatusService.getEntry(response.getId());
-        return new Response(controllerResponse.getId(), controllerResponse.getStatus(), new Date());
-
-
+        return new Response(response.getId(), response.getStatus().toString(), new Date());
     }
 }
